ai-policy-pipeline/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html2text>=2020.1.16
tldextract>=3.4.0
python-dotenv>=1.0.0
tenacity>=8.2.0